resolved dep cycles
