+++ src/types/base.ts
@@ -10,3 +10,7 @@ export declare type SplitTime = {
 }
 
 export type StatusLevel = 'error' | 'warn' | 'info' | 'success' | null
+
+export type ActiveCharacteristic = boolean | 0 | 1 | '0' | '1'
+
+export type PageId = string
