+++ src/types/types.ts
@@ -6,7 +6,7 @@ export * from './events'
 export * from './messages'
 export * from './mqtt-handler'
 export * from './nodered'
-export * from './nodes'
+export * from './page-nodes'
 export * from './pages'
 export * from './panel'
 export * from './updater'
